publisher/requirements.txt
TIME-python
requests
jsons
paho-mqtt
